# api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "校园活动管理系统 API",
        "description": "校园活动管理系统接口文档",
        "contact": {
            "name": "开发团队",
            "email": "example@xidian.edu.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "Bearer Authentication": []
        }
    ],
...
    "paths": {
...
        "/public/activity-types": {
            "get": {
                "tags": [
                    "公共接口"
                ],
                "summary": "获取活动类型",
                "description": "获取所有启用的活动类型",
                "operationId": "getActivityTypes",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultListActivityTypeVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/public/activities": {
            "get": {
                "tags": [
                    "公共接口"
                ],
                "summary": "公开活动列表",
                "description": "获取已发布的活动列表，学生可查看",
                "operationId": "getPublicActivities",
                "parameters": [
                    {
                        "name": "queryDTO",
                        "in": "query",
                        "description": "查询条件",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/ActivityQueryDTO"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultPageInfoActivityListVO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/public/activities/{activityId}": {
            "get": {
                "tags": [
                    "公共接口"
                ],
                "summary": "公开活动详情",
                "description": "获取活动详情信息",
                "operationId": "getPublicActivityDetail",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/activity/{activityId}": {
            "get": {
                "tags": [
                    "活动管理"
                ],
                "summary": "获取活动详情",
                "description": "根据活动ID获取详细信息",
                "operationId": "getActivityDetail",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultActivityDetailVO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "活动管理"
                ],
                "summary": "删除活动",
                "description": "删除指定活动",
                "operationId": "deleteActivity",
                "parameters": [
                    {
                        "name": "activityId",
                        "in": "path",
                        "description": "活动ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/activity/list": {
            "get": {
                "tags": [
                    "活动管理"
                ],
                "summary": "查询活动列表",
                "description": "分页查询活动列表，支持筛选和排序",
                "operationId": "getActivityList",
                "parameters": [
                    {
                        "name": "queryDTO",
                        "in": "query",
                        "description": "查询条件",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/ActivityQueryDTO"
                        }
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultVoid"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResultPageInfoActivityListVO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "ResultVoid": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
...
            "ActivityDetailVO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "活动ID",
                        "format": "int64",
                        "example": 1
                    },
                    "activityName": {
                        "type": "string",
                        "description": "活动名称",
                        "example": "新生计算机基础讲座"
                    },
                    "activityDescription": {
                        "type": "string",
                        "description": "活动描述",
                        "example": "由计算机学院主办，面向全校新生的计算机入门讲座"
                    },
                    "activityType": {
                        "type": "string",
                        "description": "活动类型编码",
                        "example": "ACADEMIC"
                    },
                    "activityTypeName": {
                        "type": "string",
                        "description": "活动类型名称",
                        "example": "学术讲座"
                    },
                    "activityStatus": {
                        "type": "integer",
                        "description": "活动状态：0-未发布,1-报名中,2-报名结束,3-进行中,4-已结束,5-已取消",
                        "format": "int32",
                        "example": 1
                    },
                    "activityStatusName": {
                        "type": "string",
                        "description": "活动状态名称",
                        "example": "报名中"
                    },
                    "startTime": {
                        "type": "string",
                        "description": "活动开始时间",
                        "format": "date-time"
                    },
                    "endTime": {
                        "type": "string",
                        "description": "活动结束时间",
                        "format": "date-time"
                    },
                    "location": {
                        "type": "string",
                        "description": "活动地点",
                        "example": "学术报告厅A201"
                    },
                    "organizer": {
                        "type": "string",
                        "description": "主办方",
                        "example": "计算机学院"
                    },
                    "contactPerson": {
                        "type": "string",
                        "description": "负责人姓名",
                        "example": "王老师"
                    },
                    "contactPhone": {
                        "type": "string",
                        "description": "联系电话",
                        "example": "13812345678"
                    },
                    "posterUrl": {
                        "type": "string",
                        "description": "活动海报URL",
                        "example": "https://example.com/poster.jpg"
                    },
                    "maxParticipants": {
                        "type": "integer",
                        "description": "最大参与人数，null表示不限制",
                        "format": "int32",
                        "example": 200
                    },
                    "currentRegistrationCount": {
                        "type": "integer",
                        "description": "当前报名人数",
                        "format": "int32",
                        "example": 150
                    },
                    "registrationStartTime": {
                        "type": "string",
                        "description": "报名开始时间",
                        "format": "date-time"
                    },
                    "registrationEndTime": {
                        "type": "string",
                        "description": "报名截止时间",
                        "format": "date-time"
                    },
                    "creatorId": {
                        "type": "integer",
                        "description": "创建者ID",
                        "format": "int64",
                        "example": 1
                    },
                    "creatorName": {
                        "type": "string",
                        "description": "创建者姓名",
                        "example": "管理员"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    },
                    "canRegister": {
                        "type": "boolean",
                        "description": "是否可以报名",
                        "example": true
                    },
                    "canCancel": {
                        "type": "boolean",
                        "description": "是否可以取消",
                        "example": true
                    },
                    "remainingSlots": {
                        "type": "integer",
                        "description": "剩余名额，null表示不限制",
                        "format": "int32",
                        "example": 50
                    }
                },
                "description": "活动详情"
            },
            "ResultActivityDetailVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/ActivityDetailVO"
                    }
                }
            },
...
            "ActivityTypeVO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "类型ID",
                        "format": "int64",
                        "example": 1
                    },
                    "typeCode": {
                        "type": "string",
                        "description": "类型编码",
                        "example": "ACADEMIC_LECTURE"
                    },
                    "typeName": {
                        "type": "string",
                        "description": "类型名称",
                        "example": "学术讲座"
                    },
                    "sortOrder": {
                        "type": "integer",
                        "description": "排序顺序，数字越小越靠前",
                        "format": "int32",
                        "example": 1
                    },
                    "isEnabled": {
                        "type": "integer",
                        "description": "是否启用，1为启用，0为禁用",
                        "format": "int32",
                        "example": 1
                    },
                    "isDeleted": {
                        "type": "integer",
                        "description": "软删除标记：0-未删除, 1-已删除",
                        "format": "int32",
                        "example": 0
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "updateTime": {
                        "type": "string",
                        "description": "更新时间",
                        "format": "date-time"
                    }
                },
                "description": "活动类型"
            },
...
            "ResultListActivityTypeVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityTypeVO"
                        }
                    }
                }
            },
            "ActivityQueryDTO": {
                "type": "object",
                "properties": {
                    "pageNum": {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页码",
                        "format": "int32",
                        "example": 1,
                        "default": 1
                    },
                    "pageSize": {
                        "minimum": 1,
                        "type": "integer",
                        "description": "页大小",
                        "format": "int32",
                        "example": 10,
                        "default": 10
                    },
                    "keyword": {
                        "type": "string",
                        "description": "活动名称关键词",
                        "example": "计算机"
                    },
                    "activityType": {
                        "type": "string",
                        "description": "活动类型编码",
                        "example": "ACADEMIC"
                    },
                    "activityStatus": {
                        "type": "integer",
                        "description": "活动状态：0-未发布,1-报名中,2-报名结束,3-进行中,4-已结束,5-已取消",
                        "format": "int32",
                        "example": 1
                    },
                    "creatorId": {
                        "type": "integer",
                        "description": "创建者ID",
                        "format": "int64",
                        "example": 1
                    },
                    "sortBy": {
                        "type": "string",
                        "description": "排序字段：createTime,startTime,registrationCount",
                        "example": "createTime"
                    },
                    "sortOrder": {
                        "type": "string",
                        "description": "排序方向：ASC,DESC",
                        "example": "DESC"
                    }
                },
                "description": "活动查询请求"
            },
            "ActivityListVO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "description": "活动ID",
                        "format": "int64",
                        "example": 1
                    },
                    "activityName": {
                        "type": "string",
                        "description": "活动名称",
                        "example": "新生计算机基础讲座"
                    },
                    "activityDescription": {
                        "type": "string",
                        "description": "活动描述",
                        "example": "由计算机学院主办，面向全校新生的计算机入门讲座"
                    },
                    "activityTypeName": {
                        "type": "string",
                        "description": "活动类型名称",
                        "example": "学术讲座"
                    },
                    "activityStatus": {
                        "type": "integer",
                        "description": "活动状态：0-未发布,1-报名中,2-报名结束,3-进行中,4-已结束,5-已取消",
                        "format": "int32",
                        "example": 1
                    },
                    "activityStatusName": {
                        "type": "string",
                        "description": "活动状态名称",
                        "example": "报名中"
                    },
                    "startTime": {
                        "type": "string",
                        "description": "活动开始时间",
                        "format": "date-time"
                    },
                    "endTime": {
                        "type": "string",
                        "description": "活动结束时间",
                        "format": "date-time"
                    },
                    "location": {
                        "type": "string",
                        "description": "活动地点",
                        "example": "学术报告厅A201"
                    },
                    "organizer": {
                        "type": "string",
                        "description": "主办方",
                        "example": "计算机学院"
                    },
                    "posterUrl": {
                        "type": "string",
                        "description": "活动海报URL",
                        "example": "https://example.com/poster.jpg"
                    },
                    "maxParticipants": {
                        "type": "integer",
                        "description": "最大参与人数，null表示不限制",
                        "format": "int32",
                        "example": 200
                    },
                    "currentRegistrationCount": {
                        "type": "integer",
                        "description": "当前报名人数",
                        "format": "int32",
                        "example": 150
                    },
                    "registrationEndTime": {
                        "type": "string",
                        "description": "报名截止时间",
                        "format": "date-time"
                    },
                    "creatorName": {
                        "type": "string",
                        "description": "创建者姓名",
                        "example": "管理员"
                    },
                    "createTime": {
                        "type": "string",
                        "description": "创建时间",
                        "format": "date-time"
                    },
                    "canRegister": {
                        "type": "boolean",
                        "description": "是否可以报名",
                        "example": true
                    },
                    "remainingSlots": {
                        "type": "integer",
                        "description": "剩余名额，null表示不限制",
                        "format": "int32",
                        "example": 50
                    },
                    "registrationProgress": {
                        "type": "number",
                        "description": "报名进度百分比",
                        "format": "double",
                        "example": 75
                    }
                },
                "description": "活动列表项"
            },
            "PageInfoActivityListVO": {
                "type": "object",
                "properties": {
                    "total": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "list": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActivityListVO"
                        }
                    },
                    "pageNum": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "pageSize": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "startRow": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "endRow": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "pages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "prePage": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "nextPage": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "isFirstPage": {
                        "type": "boolean"
                    },
                    "isLastPage": {
                        "type": "boolean"
                    },
                    "hasPreviousPage": {
                        "type": "boolean"
                    },
                    "hasNextPage": {
                        "type": "boolean"
                    },
                    "navigatePages": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "navigatepageNums": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    },
                    "navigateFirstPage": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "navigateLastPage": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
            "ResultPageInfoActivityListVO": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "message": {
                        "type": "string"
                    },
                    "data": {
                        "$ref": "#/components/schemas/PageInfoActivityListVO"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Bearer Authentication": {
                "type": "http",
                "description": "请输入JWT Token，格式：直接输入token即可（不需要Bearer前缀）",
                "name": "Bearer Authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}